move code extractStyle webpack to build-man
