+++ packages/build-man/src/index.ts
@@ -1,2 +1,3 @@
 export * from './sortCss';
 export * from './cssRuleObjectsToCssString';
+export * from './extractStyle';
